Dissociating objects in different files
