+++ js/components/ps-number-input.js
@@ -81,26 +81,4 @@ const PSNumberInput = function(element) {
     })
 };
 
-/**
- * Permit to init every PSInputNumber object on the page
- *
- * @param {string} selector
- * @method init
- * @returns {Object}
- */
-const PSNumbers = function(selector) {
-    this.selector = selector;
-    this.arrayInputs = [];
-
-    this.init = () => {
-        document.querySelectorAll(selector).forEach((e) => {
-            this.arrayInputs.push(new PSNumberInput(e));
-        })
-    }
-
-    this.init();
-
-    return this.arrayInputs;
-}
-
-export default PSNumbers;
+export default PSNumberInput;
